Shared Shutdown and Do_Onetime_Device_Dependent_Inits
