+++ Generals/Code/Libraries/Source/WWVegas/WW3D2/texturefilter.h
@@ -115,7 +115,8 @@ public:
 	void Set_V_Addr_Mode(TxtAddrMode mode) { VAddressMode=mode; }
 
 	// This needs to be called after device has been created
-	static void _Init_Filters(void);
+	static void _Init_Filters(TextureFilterMode texture_filter);
+	static void _Init_Filters();
 
 	static void _Set_Default_Min_Filter(FilterType filter);
 	static void _Set_Default_Mag_Filter(FilterType filter);
